:bug: mission no.3 - 참여 중인 미션 조회 방식 변경

+++ src/main/java/com/example/SpringStudy/web/dto/HomeReponseDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class HomeReponseDto {
     private String regionName;
     private Integer point;
-    private Integer completedMissionCount;
+    private Long completedMissionCount;
     private List<MyMissionResponseDto> missions;
 
     @Override
